update app sub-directories to remove v2
