--- src/applications/coronavirus-research/sign-up-v2/app-entry.jsx
@@ -1,14 +1,0 @@
-import 'platform/polyfills';
-import './sass/covid-research-volunteer.scss';
-
-import startApp from 'platform/startup';
-
-import routes from './routes';
-import reducer from './reducers';
-import manifest from './manifest.json';
-
-startApp({
-  url: manifest.rootUrl,
-  reducer,
-  routes,
-});
