--- src/applications/coronavirus-research/sign-up-v2/containers/App.jsx
@@ -1,12 +1,0 @@
-import React from 'react';
-
-import RoutedSavableApp from 'platform/forms/save-in-progress/RoutedSavableApp';
-import formConfig from '../config/form';
-
-export default function App({ location, children }) {
-  return (
-    <RoutedSavableApp formConfig={formConfig} currentLocation={location}>
-      {children}
-    </RoutedSavableApp>
-  );
-}
