--- src/applications/coronavirus-research/sign-up-v2/containers/customReviewYesNo.jsx
@@ -1,10 +1,0 @@
-import React from 'react';
-
-export default function CustomReviewYesNo({ children, uiSchema }) {
-  return (
-    <div className="review-row">
-      <dt>{uiSchema['ui:title'].props.children.props.children}</dt>
-      <dd>{children.props.formData ? 'Yes' : 'No'}</dd>
-    </div>
-  );
-}
