--- src/applications/coronavirus-research/sign-up-v2/manifest.json
@@ -1,6 +1,0 @@
-{
-  "appName": "Coronavirus Research - Volunteer - V2",
-  "entryFile": "./app-entry.jsx",
-  "entryName": "coronavirus-research-v2",
-  "rootUrl": "/coronavirus-research/volunteer/v2"
-}
