--- src/applications/coronavirus-research/sign-up-v2/sass/covid-research-volunteer.scss
@@ -1,13 +1,0 @@
-@import "~@department-of-veterans-affairs/formation/sass/shared-variables";
-@import "~@department-of-veterans-affairs/formation/sass/modules/m-process-list";
-@import "~@department-of-veterans-affairs/formation/sass/modules/m-form-process";
-@import "~@department-of-veterans-affairs/formation/sass/modules/m-progress-bar";
-@import "../../../../platform/forms/sass/m-schemaform";
-@import "~@department-of-veterans-affairs/formation/sass/modules/m-modal";
-@import "~@department-of-veterans-affairs/formation/sass/modules/m-omb-info";
-@import "../../../../platform/forms/sass/m-form-confirmation";
-
-.input-width {
-  display: inline-block;
-  max-width: units(15);
-}
