--- src/applications/coronavirus-research/sign-up-v2/tests/fixtures/reviewAndSubmitData.json
@@ -1,219 +1,0 @@
-{
-  "values": [
-    {
-      "labelId": "root_diagnosed-label",
-      "labelName": "Have you ever been diagnosed with COVID-19?",
-      "fieldId": "root_diagnosedYes",
-      "fieldName": "Yes",
-      "type": "radio"
-    },
-    {
-      "labelId": "root_DIAGNOSED_DETAILS-label",
-      "labelName": "How were you diagnosed?",
-      "reviewLabelName": "Diagnosed by:",
-      "fieldId": "root_DIAGNOSED_DETAILS_DIAGNOSED_DETAILS\\:\\:SYMPTOMS_ONLY",
-      "fieldName": "Based on my symptoms",
-      "type": "select"
-    },
-    {
-      "labelId": "root_closeContactPositive-label",
-      "labelName": "In the past month, have you been in close contact with anyone who you know tested positive for COVID-19?",
-      "fieldId": "root_closeContactPositive_2",
-      "fieldName": "I don't know",
-      "type": "radio"
-    },
-    {
-      "labelId": "root_vaccinated-label",
-      "labelName": "Have you received a COVID-19 vaccine?",
-      "fieldId": "root_vaccinatedYes",
-      "fieldName": "Yes",
-      "type": "radio"
-    },
-    {
-      "labelId": "root_VACCINATED_DETAILS-label",
-      "labelName": "Which vaccine did you receive?",
-      "fieldId": "root_VACCINATED_DETAILS_1",
-      "fieldName": "Pfizer",
-      "type": "radio"
-    },
-    {
-      "fieldId": "root_VACCINATED_DATE1Month",
-      "fieldName": "5",
-      "type": "dropDown"
-    },
-    {
-      "fieldId": "root_VACCINATED_DATE1Year",
-      "fieldName": "2021",
-      "type": "text"
-    },
-    {
-      "fieldId": "root_VACCINATED_DATE2Month",
-      "fieldName": "8",
-      "type": "dropDown"
-    },
-    {
-      "fieldId": "root_VACCINATED_DATE2Year",
-      "fieldName": "2020",
-      "type": "text"
-    },
-    {
-      "labelId": "root_VACCINATED_ADDITIONAL1-label",
-      "labelName": "Have you received any additional vaccine doses (including boosters)?",
-      "fieldId": "root_VACCINATED_ADDITIONAL1Yes",
-      "fieldName": "Yes",
-      "type": "radio"
-    },
-    {
-      "labelId": "root_VACCINATED_ADDITIONAL_DETAILS1-label",
-      "labelName": "If Yes, which vaccine did you receive?",
-      "fieldId": "root_VACCINATED_ADDITIONAL_DETAILS1_1",
-      "fieldName": "Pfizer",
-      "type": "radio"
-    },
-    {
-      "fieldId": "root_VACCINATED_ADDITIONAL_DATE1Month",
-      "fieldName": "5",
-      "type": "dropDown"
-    },
-    {
-      "fieldId": "root_VACCINATED_ADDITIONAL_DATE1Year",
-      "fieldName": "2021",
-      "type": "text"
-    },
-    {
-      "labelId": "root_hospitalized-label",
-      "labelName": "In the past 6 months, have you had to stay overnight in a hospital for treatment, care, or testing?",
-      "fieldId": "root_hospitalizedYes",
-      "fieldName": "Yes",
-      "type": "radio"
-    },
-    {
-      "labelId": "root_smokeOrVape-label",
-      "labelName": "Do you smoke or vape, or do you have a past history of smoking or vaping?",
-      "fieldId": "root_smokeOrVapeNo",
-      "fieldName": "No",
-      "type": "radio"
-    },
-    {
-      "labelId": "root_HEALTH_HISTORY-label",
-      "labelName": "Do you have a history of any of the health issues listed below?",
-      "reviewLabelName": "Health history:",
-      "fieldId": "root_HEALTH_HISTORY_HEALTH_HISTORY\\:\\:ALLERGY_VACCINE|root_HEALTH_HISTORY_HEALTH_HISTORY\\:\\:IMMUNOCOMPROMISED",
-      "fieldName": "Allergy to any vaccines|Compromised immune system (including due to HIV/AIDS)",
-      "type": "select"
-    },
-    {
-      "labelId": "root_EMPLOYMENT_STATUS-label",
-      "labelName": "Which work situation describes you?",
-      "reviewLabelName": "Employment status:",
-      "fieldId": "root_EMPLOYMENT_STATUS_EMPLOYMENT_STATUS\\:\\:EMPLOYED_HOME",
-      "fieldName": "Employed (working from home)",
-      "type": "select"
-    },
-    {
-      "labelId": "root_TRANSPORTATION-label",
-      "labelName": "How do you get to work?",
-      "reviewLabelName": "Transportation:",
-      "fieldId": "root_TRANSPORTATION_TRANSPORTATION\\:\\:CAR",
-      "fieldName": "Car",
-      "type": "select"
-    },
-    {
-      "labelId": "root_residentsInHome-label",
-      "labelName": "How many people, including you, live in your home?",
-      "fieldId": "root_residentsInHome_1",
-      "fieldName": "3 to 5",
-      "type": "radio"
-    },
-    {
-      "labelId": "root_closeContact-label",
-      "labelName": "On most days, how many people do you have close contact with outside of those who live in your home?",
-      "fieldId": "root_closeContact_1",
-      "fieldName": "1 to 10",
-      "type": "radio"
-    },
-    {
-      "labelId": "root_VETERAN-label",
-      "labelName": "Which of these best describe you?",
-      "reviewLabelName": "Relationship to the VA:",
-      "fieldId": "root_VETERAN_VETERAN\\:\\:VETERAN",
-      "fieldName": "Veteran",
-      "type": "select"
-    },
-    {
-      "labelId": "root_ELIGIBLE-label",
-      "labelName": "Are you eligible to receive care at a VA facility?",
-      "fieldId": "root_ELIGIBLENo",
-      "fieldName": "No",
-      "type": "radio"
-    },
-    {
-      "labelId": "root_FACILITY-label",
-      "labelName": "Do you receive care at a VA facility?",
-      "fieldId": "root_FACILITYNo",
-      "fieldName": "No",
-      "type": "radio"
-    },
-    {
-      "labelId": "root_GENDER-label",
-      "labelName": "What is your gender?",
-      "reviewLabelName": "Gender:",
-      "fieldId": "root_GENDER_GENDER\\:\\:FEMALE",
-      "fieldName": "Woman",
-      "type": "select"
-    },
-    {
-      "labelId": "root_RACE_ETHNICITY-label",
-      "labelName": "What is your race, ethnicity, or origin?",
-      "reviewLabelName": "Race, ethnicity, and origin:",
-      "fieldId": "root_RACE_ETHNICITY_RACE_ETHNICITY\\:\\:AMERICAN_INDIAN_ALASKA_NATIVE",
-      "fieldName": "American Indian or Alaska Native",
-      "type": "select"
-    },
-    {
-      "fieldId": "root_veteranDateOfBirthMonth",
-      "fieldName": "11",
-      "type": "dropDown"
-    },
-    {
-      "fieldId": "root_veteranDateOfBirthDay",
-      "fieldName": "22",
-      "type": "dropDown"
-    },
-    {
-      "fieldId": "root_veteranDateOfBirthYear",
-      "fieldName": "1963",
-      "type": "text"
-    },
-    {
-      "labelName": "First name",
-      "fieldId": "root_veteranFullName_first",
-      "fieldName": "fNameTest",
-      "type": "text"
-    },
-    {
-      "labelName": "Last name",
-      "fieldId": "root_veteranFullName_last",
-      "fieldName": "lNameTest",
-      "type": "text"
-    },
-    {
-      "labelName": "Email address",
-      "fieldId": "root_email",
-      "fieldName": "test@vetTest.com",
-      "type": "text"
-    },
-    {
-      "labelName": "Zip code where you currently live",
-      "fieldId": "root_zipCode",
-      "fieldName": "12345",
-      "type": "text"
-    },
-    {
-      "labelName": "Phone",
-      "fieldId": "root_phone",
-      "fieldName": "(555) 444-3333",
-      "type": "text"
-    }
-  ]
-}
