--- src/applications/coronavirus-research/sign-up-v2/tests/fixtures/toggle-covid-feature.json
@@ -1,11 +1,0 @@
-{
-  "data": {
-    "type": "feature_toggles",
-    "features": [
-      {
-        "name": "covid_volunteer_intake_enabled",
-        "value": true
-      }
-    ]
-  }
-}
