+++ src/applications/coronavirus-research/sign-up/config/temp-COVID-VACCINE-TRIAL-schema.json
