+++ src/applications/coronavirus-research/sign-up/config/va-location/DynamicRadioWidget.jsx
@@ -66,9 +66,9 @@ export function DynamicRadioWidget(props) {
           <p className="vads-u-padding-left--4 vads-u-margin-top--neg3">
             {location.attributes.name}
           </p>
-          <p className="vads-u-padding-left--4 vads-u-margin-top--neg2">{`${
-            location.attributes.city
-          } ${location.attributes.state}`}</p>
+          <p className="vads-u-padding-left--4 vads-u-margin-top--neg2">
+            {`${location.attributes.city} ${location.attributes.state}`}
+          </p>
         </>
       ),
       value: `${location.attributes.name}|${location.id}`,
