+++ src/applications/coronavirus-research/sign-up/manifest.json
