+++ src/applications/coronavirus-research/sign-up/tests/fixtures/reviewAndSubmitData.json
@@ -23,6 +23,64 @@
       "type": "radio"
     },
     {
+      "labelId": "root_vaccinated-label",
+      "labelName": "Have you received a COVID-19 vaccine?",
+      "fieldId": "root_vaccinatedYes",
+      "fieldName": "Yes",
+      "type": "radio"
+    },
+    {
+      "labelId": "root_VACCINATED_DETAILS-label",
+      "labelName": "Which vaccine did you receive?",
+      "fieldId": "root_VACCINATED_DETAILS_1",
+      "fieldName": "Pfizer",
+      "type": "radio"
+    },
+    {
+      "fieldId": "root_VACCINATED_DATE1Month",
+      "fieldName": "5",
+      "type": "dropDown"
+    },
+    {
+      "fieldId": "root_VACCINATED_DATE1Year",
+      "fieldName": "2021",
+      "type": "text"
+    },
+    {
+      "fieldId": "root_VACCINATED_DATE2Month",
+      "fieldName": "8",
+      "type": "dropDown"
+    },
+    {
+      "fieldId": "root_VACCINATED_DATE2Year",
+      "fieldName": "2020",
+      "type": "text"
+    },
+    {
+      "labelId": "root_VACCINATED_ADDITIONAL1-label",
+      "labelName": "Have you received any additional vaccine doses (including boosters)?",
+      "fieldId": "root_VACCINATED_ADDITIONAL1Yes",
+      "fieldName": "Yes",
+      "type": "radio"
+    },
+    {
+      "labelId": "root_VACCINATED_ADDITIONAL_DETAILS1-label",
+      "labelName": "If Yes, which vaccine did you receive?",
+      "fieldId": "root_VACCINATED_ADDITIONAL_DETAILS1_1",
+      "fieldName": "Pfizer",
+      "type": "radio"
+    },
+    {
+      "fieldId": "root_VACCINATED_ADDITIONAL_DATE1Month",
+      "fieldName": "5",
+      "type": "dropDown"
+    },
+    {
+      "fieldId": "root_VACCINATED_ADDITIONAL_DATE1Year",
+      "fieldName": "2021",
+      "type": "text"
+    },
+    {
       "labelId": "root_hospitalized-label",
       "labelName": "In the past 6 months, have you had to stay overnight in a hospital for treatment, care, or testing?",
       "fieldId": "root_hospitalizedYes",
@@ -44,7 +102,6 @@
       "fieldName": "Allergy to any vaccines|Compromised immune system (including due to HIV/AIDS)",
       "type": "select"
     },
-
     {
       "labelId": "root_EMPLOYMENT_STATUS-label",
       "labelName": "Which work situation describes you?",
@@ -82,6 +139,20 @@
       "fieldId": "root_VETERAN_VETERAN\\:\\:VETERAN",
       "fieldName": "Veteran",
       "type": "select"
+    },
+    {
+      "labelId": "root_ELIGIBLE-label",
+      "labelName": "Are you eligible to receive care at a VA facility?",
+      "fieldId": "root_ELIGIBLENo",
+      "fieldName": "No",
+      "type": "radio"
+    },
+    {
+      "labelId": "root_FACILITY-label",
+      "labelName": "Do you receive care at a VA facility?",
+      "fieldId": "root_FACILITYNo",
+      "fieldName": "No",
+      "type": "radio"
     },
     {
       "labelId": "root_GENDER-label",
