+++ src/applications/coronavirus-research/sign-up/tests/reviewAndSubmit.cypress.spec.js
@@ -2,7 +2,7 @@ import path from 'path';
 import featureTogglesEnabled from './fixtures/toggle-covid-feature.json';
 
 describe('COVID-19 Research Form', () => {
-  describe('when entering valid information and submitting', () => {
+  describe('when entering valid information and submitting the sign up form', () => {
     before(() => {
       cy.intercept('GET', '/v0/feature_toggles*', featureTogglesEnabled).as(
         'feature',
@@ -13,7 +13,6 @@ describe('COVID-19 Research Form', () => {
       cy.visit('coronavirus-research/volunteer/sign-up');
       cy.injectAxe();
     });
-
     it('should load form page', () => {
       cy.url().should('include', 'coronavirus-research/volunteer/sign-up');
       cy.axeCheck();
