+++ src/applications/coronavirus-research/update/tests/reviewAndSubmitUpdate.cypress.spec.js
@@ -14,7 +14,7 @@ describe('COVID-19 Research Form', () => {
       cy.injectAxe();
     });
     // Skipping tests until routing is in place
-    it.skip('should load form page', () => {
+    it('should load form page', () => {
       cy.url().should('include', 'coronavirus-research/volunteer/update');
       cy.axeCheck();
       cy.get('h1').contains(
@@ -23,7 +23,7 @@ describe('COVID-19 Research Form', () => {
       cy.axeCheck();
     });
 
-    it.skip('should successfully submit the Covid Research form', () => {
+    it('should successfully submit the Covid Research form', () => {
       cy.fixture(
         path.join(__dirname, 'fixtures', 'reviewAndSubmitData.json'),
       ).then(dataElements => {
